Moved interpolation to be a component part of the engine, made the window class be seperated by namespace so both GLFW and SDL get to keep the same class name of Window
