+++ 3Sisters-Engine/src/ecs/systems/ecs_physics.cpp
@@ -2,7 +2,6 @@
 
 // include Sprite Renderer
 #include "ecs/components/transform.hpp"
-#include "engine/interpolation.hpp"
 #include "engine/physics.hpp"
 #include "engine/sprite_renderer.hpp"
