+++ 3Sisters-Engine/src/window/glfw_window.cpp
@@ -1,5 +1,9 @@
 #include <window/glfw_window.hpp>
 
+// include the namespace
+using namespace GLFW;
+
+// include standard libraries
 #include <iostream>
 #include <chrono>
