+++ 3Sisters-Engine/src/window/sdl_window.cpp
@@ -1,5 +1,7 @@
 #include <window/sdl_window.hpp>
 
+using namespace SDL;
+
 Window::Window(){
 
 }
